Make `iced_native` subscription input opaque
